+++ src/main/java/pixelitor/filters/impl/AngularWavesFilter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Laszlo Balazs-Csiki and Contributors
+ * Copyright 2022 Laszlo Balazs-Csiki and Contributors
  *
  * This file is part of Pixelitor. Pixelitor is free software: you
  * can redistribute it and/or modify it under the terms of the GNU
@@ -26,9 +26,9 @@ import pixelitor.filters.AngularWaves;
  */
 public class AngularWavesFilter extends CenteredTransformFilter {
     private double radialWL; // Radial Wavelength
-    private float phase;
-    private float zoom;
-    private float amount;
+    private double phase;
+    private double zoom;
+    private double amount;
     private int waveType;
 
     public AngularWavesFilter() {
@@ -37,8 +37,8 @@ public class AngularWavesFilter extends CenteredTransformFilter {
 
     @Override
     protected void transformInverse(int x, int y, float[] out) {
-        float dx = x - cx;
-        float dy = y - cy;
+        double dx = x - cx;
+        double dy = y - cy;
         double r = Math.sqrt(dx * dx + dy * dy);
         double angle = FastMath.atan2(dy, dx);
 
@@ -49,26 +49,26 @@ public class AngularWavesFilter extends CenteredTransformFilter {
         angle += fa * amount;
 
         double zoomedR = r / zoom;
-        float u = (float) (zoomedR * FastMath.cos(angle));
-        float v = (float) (zoomedR * FastMath.sin(angle));
+        double u = zoomedR * FastMath.cos(angle);
+        double v = zoomedR * FastMath.sin(angle);
 
-        out[0] = u + cx;
-        out[1] = v + cy;
+        out[0] = (float) (u + cx);
+        out[1] = (float) (v + cy);
     }
 
     public void setRadialWL(double radialWL) {
         this.radialWL = radialWL;
     }
 
-    public void setPhase(float phase) {
+    public void setPhase(double phase) {
         this.phase = phase;
     }
 
-    public void setZoom(float zoom) {
+    public void setZoom(double zoom) {
         this.zoom = zoom;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
